clean up external commands to be closer to FKZ ones
